continue moving params to config files
