+++ sramgen/src/config/inv_chain.rs
@@ -1,0 +1,11 @@
+pub struct InvChainParams<'a> {
+    pub prefix: &'a str,
+    pub num: usize,
+}
+
+#[derive(Copy, Clone, Eq, PartialEq, Hash, Debug)]
+pub struct InvChainGridParams<'a> {
+    pub prefix: &'a str,
+    pub rows: usize,
+    pub cols: usize,
+}
