+++ sramgen/src/config/sram.rs
@@ -1,4 +1,23 @@
-use crate::config::ControlMode;
+use anyhow::Result;
+use serde::{Deserialize, Serialize};
+use std::fs;
+use std::path::Path;
+
+#[derive(Debug, Eq, PartialEq, Clone, Hash, Serialize, Deserialize)]
+pub struct SramConfig {
+    pub num_words: i32,
+    pub data_width: i32,
+    pub mux_ratio: i32,
+    pub write_size: i32,
+    pub control: ControlMode,
+}
+
+#[derive(Debug, Eq, PartialEq, Copy, Clone, Hash, Serialize, Deserialize)]
+pub enum ControlMode {
+    Simple,
+    SimpleChipSelect,
+    Replica,
+}
 
 #[derive(Debug)]
 pub struct SramParams {
@@ -21,4 +40,10 @@ pub struct SramParams {
     pub addr_width: usize,
 
     pub control: ControlMode,
+}
+
+pub fn parse_sram_config(path: impl AsRef<Path>) -> Result<SramConfig> {
+    let contents = fs::read_to_string(path)?;
+    let data = toml::from_str(&contents)?;
+    Ok(data)
 }
