+++ sramgen/src/config/wmask_control.rs
@@ -1,0 +1,6 @@
+#[derive(Debug, Clone)]
+pub struct WriteMaskControlParams {
+    pub name: String,
+    pub width: i64,
+    pub and_params: AndParams,
+}
