+++ sramgen/src/plan/mod.rs
@@ -1,5 +1,5 @@
 use crate::config::sram::SramParams;
-use crate::config::{ControlMode, SramConfig};
+use crate::config::sram::{ControlMode, SramConfig};
 use crate::layout::sram::draw_sram;
 use crate::paths::{out_bin, out_gds, out_sram, out_verilog};
 use crate::plan::extract::ExtractionResult;
